backend/requirements.txt
Flask
flask-cors
transformers
torch
PyMuPDF
python-docx
python-pptx
pytesseract
Pillow
textract
pandas
openpyxl
